python/test-repo/pyproject.toml
[tool.poetry]
name = "popbot-service"
version = "1.0.0"
description = "PopBot Service by Populix Devs"
authors = ["developer@populix.co"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.13"
fastapi = "0.110.0"
uvicorn = "0.27.1"
boto3 = "1.34.133"
httpx = "0.27.0"
langchain = "0.2.3"
langchain-google-vertexai = "1.0.5"
langsmith = "0.1.76"
pandas = "2.2.2"
python-dotenv = "1.0.1"
python-logging-loki = "0.3.1"
openpyxl = "3.1.4"
psycopg2-binary = "2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"
coverage = "^7.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
